Remove compilation warnings from llvm-3.5svn.

+++ source/boundaries/boxes.hpp
@@ -98,7 +98,7 @@ public:
 		}
 
 		bool overlaps = true;
-		for (size_t i = 0; i < cell_start.size(); i++) {
+		for (size_t i = 0; i < size_t(cell_start.size()); i++) {
 			if (
 				cell_start[i] >= this->ends[index][i]
 				or cell_end[i] <= this->starts[index][i]
